✨ improve source string
